feat: remove local apiconnect module to use publish one
- Add apiconnect utils and worker
- Remove local apiconnect
- Update the function to pass additional header
