+++ package.json
@@ -80,6 +80,8 @@
         "typescript-eslint": "^8.15.0"
     },
     "dependencies": {
+        "@dashlane/apiconnect-utils": "^3.1.9",
+        "@dashlane/apiconnect-worker": "^3.1.9",
         "@dashlane/nsm-attestation": "^1.0.2",
         "@inquirer/prompts": "^7.1.0",
         "@json2csv/plainjs": "^7.0.6",
