--- src/modules/api-connect/index.ts
@@ -1,3 +1,0 @@
-export * from './postRequestAPI.js';
-export * from './signRequest.js';
-export * from './types.js';
